requirements.txt
asgiref==3.5.0
Django==4.0.4
Pillow==9.1.0
sqlparse==0.4.2
django-admin-sortable2==2.0.3
django-tinymce==3.4.0
django-environ==0.8.1


